// packages/measurements/src/measure.types.ts
interface PolylineSegment {
  show: boolean
  positions: Array<Cesium.Cartesian3>
  drawStatus: number
}

interface DistanceMeasurementPolylineSegment extends PolylineSegment {
  distance: number
  labelPosition: Cesium.Cartesian3
  xyPolylinePositions?: Array<Cesium.Cartesian3>
  xyBoxPositions?: Array<Cesium.Cartesian3>
  xDistance?: number
  xLabelPosition?: Cesium.Cartesian3
  yDistance?: number
  yLabelPosition?: Cesium.Cartesian3
  xAngle?: number
  xAnglePosition?: Cesium.Cartesian3
  yAngle?: number
  yAnglePosition?: Cesium.Cartesian3
}